Add experimental relationship query branch
This is fully functional, but the code is not tidy at all. These needs
a review, and is likely to need a lot of revision to clean it up, and
hopefully simplify it.

Also needs README to be modified to provide info on usage of new
functionality

Allows both kinds of relationship query:
 - Inner queries (Select *, (Select * from rel__r) from obj__c)
 - Dot queries (Select *, rel__r.field__c from obj__c)